SSAFY/django/online_practice/10_16/online_practice/django_ws_10_1/requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.15.2
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
